refactor(engine): extract inventory utils; remove CLI entry points
Phase 0 of terminal-mode removal:
- Add src/inventory_utils.py with get_gold/transfer_gold/transfer_item
  (extracted from interface.py; error paths now log instead of print)
- interface.py re-exports these for its remaining terminal menu classes
- Repoint API (game_service), events.py, objects.py to inventory_utils
- Delete CLI entry points: game.py, intro_scene.py, open_terminal.py
- Repoint affected test patch targets to inventory_utils
