+++ tests/test_events.py
@@ -297,7 +297,7 @@ class TestLootEvent:
         assert event.input_options[3]["value"] == "exit"
 
     @patch('neotermcolor.cprint')
-    @patch('interface.transfer_item')
+    @patch('inventory_utils.transfer_item')
     def test_loot_event_process_take_all(self, mock_transfer_item, mock_cprint):
         """Test LootEvent process with 'all' input."""
         mock_container = MagicMock()
@@ -321,7 +321,7 @@ class TestLootEvent:
         assert result == {"success": True}
 
     @patch('neotermcolor.cprint')
-    @patch('interface.transfer_item')
+    @patch('inventory_utils.transfer_item')
     def test_loot_event_process_take_specific_item(self, mock_transfer_item, mock_cprint):
         """Test LootEvent process with specific item index."""
         mock_container = MagicMock()
