+++ tests/test_objects_coverage_boost.py
@@ -229,7 +229,7 @@ class TestContainer:
         gold = items.Gold(5)
         c.inventory = [gold]
         with (
-            patch("interface.transfer_item") as mock_transfer,
+            patch("inventory_utils.transfer_item") as mock_transfer,
             patch("builtins.print"),
             patch.object(c, "refresh_description"),
             patch.object(c, "process_events"),
