+++ tests/test_objects_positions_tier3.py
@@ -1893,7 +1893,7 @@ class TestMoreContainer:
         item2.description = "Item 2 description"
         container = Container(player=player, tile=tile, start_open=True, inventory=[item1, item2])
         # transfer_item is imported inside the method, so patch it there
-        with patch("interface.transfer_item"):
+        with patch("inventory_utils.transfer_item"):
             container.take_all(player)
             # If we got here without error, the test passed
             assert len(container.inventory) == 0 or len(container.inventory) == 2
